requirements.txt
asgiref==3.2.10
Django==3.1.1
django-multipleselectfield==0.2.0.post2
django-multiselectfield==0.1.12
django-smart-selects==1.5.9
djangorestframework==3.11.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
